% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_m5_rmse}
\alias{eval_m5_rmse}
\title{Dygraph plot}
\usage{
eval_m5_rmse(pred, real, id_vars)
}
\arguments{
\item{pred}{}

\item{real}{}
}
\value{
None
}
\description{
Creates a TS dygraph plot for the products in ids
}
