% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prc_ggl_data}
\alias{prc_ggl_data}
\title{Process Google Analytics Data}
\usage{
prc_ggl_data(raw_data, cols_to_choose = prep_cols())
}
\arguments{
\item{raw_data}{raw Google Analytics data}

\item{cols_to_choose}{columns to choose}
}
\value{
processed Google Analytics data
}
\description{
Process Google Analytics Data
}
