% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prep_cols}
\alias{prep_cols}
\title{Prepare columns for data}
\usage{
prep_cols(package = "GglAnalyticsRev")
}
\arguments{
\item{package}{name of package}
}
\value{
column names to choose
}
\description{
Prepare columns for data
}
