% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prep_partition_train}
\alias{prep_partition_train}
\title{Prepare Partitioned Train Sets}
\usage{
prep_partition_train(df, k)
}
\arguments{
\item{df}{data for partitioning}

\item{k}{the kth partition}
}
\value{
the kth processed training partition
}
\description{
Prepare Partitioned Train Sets
}
