% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_train_test_temporally}
\alias{split_train_test_temporally}
\title{Split train test temproally}
\usage{
split_train_test_temporally(train, id_vars, day_vars)
}
\arguments{
\item{train}{}

\item{day_vars}{}
}
\value{
None
}
\description{
Take the last 28 days from train out
}
